Multiple changes made and bugs fixed. Eval circuit func now a pure func and not a member of class circuit. Adde dmore func parameters to run genetic algorithm member func. Funcs declared in h files tsp_funcs.h

+++ uav/postal_region.h
@@ -4,7 +4,7 @@ using namespace std;
 struct address_metadata
 {
 	string post_code;
-	double x_coord = 0, y_coord = 0;
+	float x_coord = 0, y_coord = 0;
 	string out_code;
 	string in_code;
 	int id = -1;
